Move heartbeat-related tests in test_storages.py to another file
